+++ changelog.md
@@ -3,8 +3,10 @@
 This project adheres to semantic versioning and all major changes will
 be recorded in this file.
 
-## [unreleased]
+## [0.3.0] 2023-07-16
 
+- Add option --append-file to mkget
+- Rename mkgetters to mkget
 - Add mkset
 
 ## [0.2.0] 2023-07-15